+++ RELEASE-NOTES.md
@@ -43,6 +43,7 @@ The Axom project release numbers follow [Semantic Versioning](http://semver.org/
 - Conduit version changed to v. 0.7.2
 - Renames `AXOM_DEBUG_VAR` macro to `AXOM_UNUSED_VAR` since there were many cases where the latter
   was not the appropriate name. This macro elides warnings about unused variables
+- Inlet's `isUserProvided` can now be used to query the status of subobjects of a `Container` via a name parameter
 
 ### Fixed
 - Fixed Primal's `intersect(Ray, Segment)` calculation for Segments that do not have unit length